Changed code in Controller to call methods from mock repo, run app and test get all and get by id using swagger

+++ Product/Data/IProizvodRepozitorijum.cs
@@ -9,6 +9,6 @@ namespace Product.Data
      public interface IProizvodRepozitorijum
     {
         IEnumerable<Proizvod> VratiProizvode();
-        Proizvod VratiProizvodPoId(int id);
+        Proizvod VratiProizvodPoId(long id);
     }
 }
